ignore and not log warn when sysvar setting match
